+++ wcomponents-examples/src/main/java/com/github/bordertech/wcomponents/examples/theme/ajax/AjaxWPanelExample.java
@@ -17,8 +17,8 @@ import com.github.bordertech.wcomponents.subordinate.Show;
 import com.github.bordertech.wcomponents.subordinate.WSubordinateControl;
 
 /**
- * This example demonstrates a {@link WPanel} of type {@link WPanel.Type.LAZY} that loads its content via AJAX when made
- * visible.
+ * This example demonstrates a {@link WPanel} of type {@link WPanel.PanelMode} LAZY that loads its content via AJAX when
+ * made visible.
  * <p>
  * It uses a {@link WSubordinateControl} on a {@link WRadioButtonSelect} to toggle the visibility of the {@link WPanel}.
  * </p>